trading-bot/requirements.txt
# Core trading dependencies
fyers-apiv3==3.0.6
pandas==2.1.4
numpy==1.24.3

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2

# Technical Analysis
ta==0.10.2

# API and Data
requests==2.31.0
python-dotenv==1.0.0

# GUI Framework
tkinter  # Built-in with Python

# Utilities
pytz==2023.3
datetime  # Built-in with Python

# Development and Testing
pytest==7.4.3
pytest-cov==4.1.0

# Logging
logging  # Built-in with Python
